Add lightweight Arduino_GFX lib
Arduino_GFX includes tons of code for various displays. This results in increased build time and side
and compatibility errors. All of that drivers are useless in this project, only Arduino canvas is used.
Excluding all that code and bundling a Arduino_GFX lib in the project elliminates that useless burden.